Cleaned up commit.
The problem was thinking about Yaw, as in "turning the camera around". But that's not how it worked. Essentially, I was turning the camera around, and then worrying about its location. But, since I can only affect rotationally applied "swivel" values, this constricted me. I was stuck on it.

The solution was to think about how Pitch affects the swivel and the camera. So, if Pitch "travels" in a circle around the location (pi radians or 32767.0f), it will be on the opposite side, but upside down. Therefore the Roll should turn the camera upside-right by pi radians or 32767.0f.